move enrich_qsos to sota_api.py
